Fix search window used in paginated OTP 2 calls
Now the search will use the calculated search window from the
initial search in the following OTP 2 calls. This is the
expected way to use the API.

If the search window is not passed a long to the next calls,
a new one is calculated every time, and a consequence is that not
all results are returned.