Limit consecutive calls to JourneyPlanner to a max of 15
If JourneyPlanner keeps returning no results, but metadata with a
small time offset, we would just continue calling the API until 7
days ahead of search time was reached. If every call is just one
hour different than the one before, this means a lot of calls, and
Google App Engine times out. A limit of 15 consecutive, resultless
calls prevents this.

Searches where 'from' and 'to' are very close could trigger this.