Remove OTP 1 search logic
We're all aboard OTP 2 now (except for replace-leg) and it is time
to start the clean-up. This removes most of the OTP 1 logic.

The SDK has been updated in order to use generated types for OTP 2.
There is still some work to do to get rid of the old SDK types for
Leg, LegMode and such, but I'm saving that for another PR.